+++ R/tm_g_distribution.R
@@ -174,7 +174,7 @@ ui_distribution <- function(id, ...) {
         tabPanel("Histogram", plot_with_settings_ui(id = ns("hist_plot"))),
         tabPanel("QQplot", plot_with_settings_ui(id = ns("qq_plot")))
       ),
-      h3("Summary statistics"),
+      h3("Statistics Table"),
       DT::dataTableOutput(ns("summary_table")),
       h3("Tests"),
       DT::dataTableOutput(ns("t_stats"))
@@ -289,12 +289,7 @@ ui_distribution <- function(id, ...) {
       ),
       panel_item(
         "Statistics Table",
-        sliderInput(ns("roundn"), "Round to n digits", min = 0, max = 10, value = 2),
-        checkboxInput(
-          ns("add_stats_plot"),
-          label = "Overlay params table",
-          value = TRUE
-        )
+        sliderInput(ns("roundn"), "Round to n digits", min = 0, max = 10, value = 2)
       ),
       panel_item(
         title = "Plot settings",
@@ -349,7 +344,7 @@ srv_distribution <- function(input,
                      )
                    }
                    ANL <- datasets$get_data(as.character(dist_var[[1]]$dataname), filtered = TRUE) # nolint
-                   params <- get_dist_params(ANL[[dist_var2]], input$t_dist)
+                   params <- get_dist_params(as.numeric(na.omit(ANL[[dist_var2]])), input$t_dist)
                    params_vec <- round(unname(unlist(params)), 2)
                    params_names <- names(params)
 
@@ -416,8 +411,8 @@ srv_distribution <- function(input,
         )
       )
       common_stack_push(substitute(
-        expr = ANL <- ANL %>% dplyr::mutate(g_var_name := forcats::fct_explicit_na(as.factor(g_var_name))), # nolint
-        env = list(g_var_name = g_var_name)
+        expr = ANL[[g_var]] <- forcats::fct_explicit_na(as.factor(ANL[[g_var]]), "NA"), # nolint
+        env = list(g_var = g_var)
       ))
     }
 
@@ -429,8 +424,8 @@ srv_distribution <- function(input,
         )
       )
       common_stack_push(substitute(
-        expr = ANL <- ANL %>% dplyr::mutate(s_var_name := forcats::fct_explicit_na(as.factor(s_var_name))), # nolint
-        env = list(s_var_name = s_var_name)
+        expr = ANL[[s_var]] <- forcats::fct_explicit_na(as.factor(ANL[[s_var]]), "NA"), # nolint
+        env = list(s_var = s_var)
       ))
     }
 
@@ -518,7 +513,6 @@ srv_distribution <- function(input,
   dist_plot_r_chunks <- eventReactive(
     eventExpr = {
       common_code_chunks()
-      input$add_stats_plot
       input$scales_type
       input$main_type
       input$bins
@@ -552,7 +546,6 @@ srv_distribution <- function(input,
       dist_param1 <- input$dist_param1
       dist_param2 <- input$dist_param2
 
-      add_stats_plot <- input$add_stats_plot # nolint
       scales_type <- input$scales_type
       ndensity <- 512
       main_type_var <- input$main_type
@@ -593,26 +586,28 @@ srv_distribution <- function(input,
         )
       } else if (length(s_var) == 0 && length(g_var) != 0) {
         substitute(
-          expr = ggplot(ANL, aes(dist_var_name)) +
+          expr = ggplot(ANL[ANL[[g_var]] != "NA", ], aes(dist_var_name)) +
             geom_histogram(position = "identity", aes_string(y = m_type), bins = bins_var, alpha = 0.3) +
             facet_wrap(~g_var_name, ncol = 1, scales = scales_raw),
           env = list(
             m_type = m_type,
             bins_var = bins_var,
             dist_var_name = dist_var_name,
+            g_var = g_var,
             g_var_name = g_var_name,
             scales_raw = tolower(scales_type)
           )
         )
       } else {
         substitute(
-          expr = ggplot(ANL, aes(dist_var_name, col = s_var_name)) +
+          expr = ggplot(ANL[ANL[[g_var]] != "NA", ], aes(dist_var_name, col = s_var_name)) +
             geom_histogram(position = "identity", aes_string(y = m_type, fill = s_var), bins = bins_var, alpha = 0.3) +
             facet_wrap(~g_var_name, ncol = 1, scales = scales_raw),
           env = list(
             m_type = m_type,
             bins_var = bins_var,
             dist_var_name = dist_var_name,
+            g_var = g_var,
             s_var = s_var,
             g_var_name = g_var_name,
             s_var_name = s_var_name,
@@ -640,34 +635,11 @@ srv_distribution <- function(input,
         )
       }
 
-      if (add_stats_plot) {
-        datas <- if (length(t_dist) != 0 && m_type == "..density.." && length(g_var) == 0 && length(s_var) == 0) {
-          distplot_stack_push(substitute(
-            expr = {
-              df_params <- as.data.frame(append(params, list(name = t_dist)))
-            },
-            env = list(t_dist = t_dist)
-          ))
-
-          quote(data.frame(
-            x = c(0.7, 0), y = c(1, 1),
-            tb = I(c(list(df_params = df_params), list(summary_table = summary_table)))
-          ))
-        } else {
-          quote(data.frame(
-            x = 0, y = 1,
-            tb = I(list(summary_table = summary_table))
-          ))
-        }
-
-        label <- if (!is_empty(g_var)) {
-          substitute(
-            expr = split(tb$summary_table, tb$summary_table$g_var_name, drop = TRUE),
-            env = list(g_var = g_var, g_var_name = g_var_name)
-          )
-        } else {
-          substitute(expr = tb, env = list())
-        }
+      if (length(t_dist) != 0 && m_type == "..density.." && length(g_var) == 0 && length(s_var) == 0) {
+        distplot_stack_push(substitute(df_params <- as.data.frame(append(params, list(name = t_dist))),
+                                       env = list(t_dist = t_dist)))
+        datas <- quote(data.frame(x = 0.7, y = 1, tb = I(list(df_params = df_params))))
+        label <- quote(tb)
 
         plot_call <- substitute(
           expr = plot_call + ggpp::geom_table_npc(
@@ -733,7 +705,6 @@ srv_distribution <- function(input,
     eventExpr = {
       common_code_chunks()
       input$scales_type
-      input$add_stats_plot
       input$qq_line
       is.null(input$ggtheme)
     },
@@ -765,7 +736,6 @@ srv_distribution <- function(input,
       dist_param2 <- input$dist_param2
 
       scales_type <- input$scales_type
-      add_stats_plot <- input$add_stats_plot
       ggtheme <- input$ggtheme
 
       validate(need(ggtheme, "Please select a theme."))
@@ -789,20 +759,22 @@ srv_distribution <- function(input,
         )
       } else if (length(s_var) == 0 && length(g_var) != 0) {
         substitute(
-          expr = ggplot(ANL, aes_string(sample = dist_var)) +
+          expr = ggplot(ANL[ANL[[g_var]] != "NA", ], aes_string(sample = dist_var)) +
             facet_wrap(~g_var_name, ncol = 1, scales = scales_raw),
           env = list(
             dist_var = dist_var,
+            g_var = g_var,
             g_var_name = g_var_name,
             scales_raw = tolower(scales_type)
           )
         )
       } else {
         substitute(
-          expr = ggplot(ANL, aes_string(sample = dist_var, color = s_var)) +
+          expr = ggplot(ANL[ANL[[g_var]] != "NA", ], aes_string(sample = dist_var, color = s_var)) +
             facet_wrap(~g_var_name, ncol = 1, scales = scales_raw),
           env = list(
             dist_var = dist_var,
+            g_var = g_var,
             s_var = s_var,
             g_var_name = g_var_name,
             scales_raw = tolower(scales_type)
@@ -821,37 +793,11 @@ srv_distribution <- function(input,
         env = list(plot_call = plot_call, mapped_dist = unname(map_dist[t_dist]))
       )
 
-      if (add_stats_plot) {
-        datas <- if (length(t_dist) != 0 && length(g_var) == 0 && length(s_var) == 0) {
-          qqplot_stack_push(substitute(
-            expr = {
-              df_params <- as.data.frame(append(params, list(name = t_dist)))
-            },
-            env = list(t_dist = t_dist)
-          ))
-
-          quote(data.frame(
-            x = c(0.7, 0), y = c(1, 1),
-            tb = I(c(
-              list(df_params = df_params),
-              list(summary_table = summary_table)
-            ))
-          ))
-        } else {
-          quote(data.frame(
-            x = 0, y = 1,
-            tb = I(list(summary_table = summary_table))
-          ))
-        }
-
-        label <- if (!is_empty(g_var)) {
-          substitute(
-            expr = split(tb$summary_table, tb$summary_table$g_var_name, drop = TRUE),
-            env = list(g_var = g_var, g_var_name = g_var_name)
-          )
-        } else {
-          substitute(expr = tb, env = list())
-        }
+      if (length(t_dist) != 0 && length(g_var) == 0 && length(s_var) == 0) {
+        qqplot_stack_push(substitute(df_params <- as.data.frame(append(params, list(name = t_dist))),
+                                     env = list(t_dist = t_dist)))
+        datas <- quote(data.frame(x = 0.7, y = 1, tb = I(list(df_params = df_params))))
+        label <- quote(tb)
 
         plot_call <- substitute(
           expr = plot_call +